Cambio en alta de empleado debido a que si borro uno me pisaba el id

+++ TP3/Employee.c
@@ -144,10 +144,10 @@ int employee_setSueldo(Employee* this,int sueldo){
 
 
 }
-float employee_getSueldo(Employee* this){
+int employee_getSueldo(Employee* this){
 
 
-   float sueldo;
+    int sueldo;
     if(this!= NULL )
     {
         sueldo = this->sueldo;
